Renamed all.py to common.py
